services/api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
paho-mqtt==2.1.0
websockets==12.0
pydantic-settings==2.4.0
httpx==0.27.0
jinja2==3.1.4
weasyprint==62.3
pydyf==0.10.0
